undo deletion, search query

+++ app/src/main/java/com/example/room/fragments/list/ListFragment.kt
@@ -5,6 +5,7 @@ import android.view.*
 import androidx.fragment.app.Fragment
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.widget.SearchView
 import androidx.core.view.isVisible
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -13,13 +14,15 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.room.R
 import com.example.room.databinding.FragmentListBinding
 import com.example.room.model.User
+import com.example.room.util.setupToast
 import com.example.room.viewmodel.UserViewModel
+import com.google.android.material.snackbar.Snackbar
 
 
-class ListFragment : Fragment(), ListAdapter.ItemClickListener {
+class ListFragment : Fragment(), ListAdapter.ItemClickListener, SearchView.OnQueryTextListener{
 
-    private  var _binding: FragmentListBinding? = null
-    private val binding  get() = _binding!!
+    private var _binding: FragmentListBinding? = null
+    private val binding get() = _binding!!
 
     private lateinit var userViewModel: UserViewModel
     private lateinit var myAdapter: ListAdapter
@@ -35,19 +38,17 @@ class ListFragment : Fragment(), ListAdapter.ItemClickListener {
             findNavController().navigate(R.id.action_listFragment_to_addFragment)
         }
 
+
+        //show the loading text as the data is retrieved
         binding.progressBar.isVisible = true
         binding.textViewLoading.isVisible = true
 
         //RecyclerView
         myAdapter = ListAdapter(this)
-//        myAdapter.itemClickListener = { view,item, position ->
-//            if (view.id == R.id.btnDelete){
-//                Toast.makeText(requireContext().applicationContext, "clicked ${R.id.btnDelete}", Toast.LENGTH_SHORT).show()
-//            }
-//        }
         binding.myRecyclerView.apply {
             adapter = myAdapter
             layoutManager = LinearLayoutManager(requireContext())
+
         }
 
         //user viewModel
@@ -57,6 +58,13 @@ class ListFragment : Fragment(), ListAdapter.ItemClickListener {
             binding.textViewLoading.isVisible = false
 
             myAdapter.setData(user)
+
+            if (user.isEmpty()){
+                binding.tvNoTask.isVisible = true
+            }else if (user.isNotEmpty()){
+                binding.tvNoTask.isVisible = false
+            }
+
         })
 
         //setup the delete menu
@@ -65,29 +73,55 @@ class ListFragment : Fragment(), ListAdapter.ItemClickListener {
         return binding.root
     }
 
+
     //inflate the delete menu
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.delete_menu, menu)
+        inflater.inflate(R.menu.my_main_menu, menu)
+
+        val search = menu.findItem(R.id.menu_search)
+        val searchView =  search?.actionView as SearchView
+        searchView.isSubmitButtonEnabled = true
+        searchView.setOnQueryTextListener(this)
+    }
+
+
+    //it will be triggered when one character is typed
+    override fun onQueryTextChange(query: String?): Boolean {
+        if (query != null){
+            searchDatabase(query)
+        }
+        return true
+    }
+
+    //it will be triggered when a character(s) is submitted only
+    override fun onQueryTextSubmit(query: String?): Boolean {
+        if (query != null){
+            searchDatabase(query)
+        }
+        return true
+    }
+
+    //it fetches data from the dataBase
+    private fun searchDatabase(search: String){
+
+        //the searchQuery should be like this so that our sql can understand
+        val searchQuery = "%$search%"
+
+        //we observe the database to see if it has the typed text and return it
+        userViewModel.searchDatabase(searchQuery).observe(viewLifecycleOwner, Observer { user ->
+            user.let {
+                myAdapter.setData(it)
+            }
+        })
     }
 
     //set item click listener to my delete menu items
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when(item.itemId){
-            R.id.delete ->{
-                userViewModel.readAllData.observe(this, Observer {user ->
-                    if (!user.isEmpty()){
-                        //the database contains user(even a single one) delete all
-                        deleteAllUsers()
-                    }else{
-                        val builder = AlertDialog.Builder(requireContext())
-                        builder.setNegativeButton("Ok"){_,_ ->
-
-                        }
-                        builder.setTitle("Sorry!")
-                        builder.setMessage("We couldn't find any data to delete")
-                        builder.create().show()
-                    }
-                })
+        when (item.itemId) {
+            R.id.menu_delete -> {
+                deleteAllUsers()
+            }
+            else -> {
             }
         }
         return true
@@ -96,19 +130,14 @@ class ListFragment : Fragment(), ListAdapter.ItemClickListener {
 
     //clickListeners of my recyclerView items
     override fun onItemClicked(view: View, user: User, position: Int) {
-
         when (view.id) {
             R.id.btnDelete -> {
                 deleteUser(user)
             }
-            R.id.btnEdit -> {
+            else -> {
+                //root layout clickListener //to update fragment
                 val action = ListFragmentDirections.actionListFragmentToUpdateFragment(user)
                 findNavController().navigate(action)
-                //Toast.makeText(requireContext().applicationContext, "clicked the edit btn of user: of user: ${user.firstName}", Toast.LENGTH_SHORT).show()
-            }
-            else -> {
-                //root layout clickListener
-                //Toast.makeText(requireContext().applicationContext, "clicked the root item of user: ${user.firstName}", Toast.LENGTH_SHORT).show()
             }
         }
 
@@ -116,39 +145,78 @@ class ListFragment : Fragment(), ListAdapter.ItemClickListener {
     }
 
     //delete a single user
-    private fun deleteUser(user:User){
+    private fun deleteUser(user: User) {
         val builder = AlertDialog.Builder(requireContext())
-        builder.setPositiveButton("Yes") {_, _, ->
+        builder.setPositiveButton("Yes") { _, _ ->
+
             userViewModel.deleteUser(user)
-            Toast.makeText(requireContext().applicationContext, "user ${user.firstName} successfully deleted", Toast.LENGTH_SHORT).show()
+            snackBarToUndoDeletedUser(user,binding.root)
         }
-        builder.setNegativeButton("No"){_,_ ->
+        builder.setNegativeButton("No") { _, _ ->
 
         }
-        builder.setTitle("Delete ${user.firstName}?")
-        builder.setMessage("Do you want to delete user ${user.firstName}?")
+        builder.setTitle("Delete ${user.title}?")
+        builder.setMessage("Do you want to delete ${user.title}?")
         builder.create()
         builder.show()
     }
 
+    //Snack Bar to undo deleted user
+    private fun snackBarToUndoDeletedUser(user: User, view: View){
+        val snackbar = Snackbar.make(view, "You deleted an item.", Snackbar.LENGTH_LONG)
+        snackbar.setAction("Undo") { _, ->
+            userViewModel.addUser(user)
+            myAdapter.notifyDataSetChanged()
+
+        }
+        snackbar.show()
+    }
+
+
 
     //delete all users
-    private fun deleteAllUsers(){
-        val builder = AlertDialog.Builder(requireContext())
-        builder.setPositiveButton("Yes"){_,_ ->
-            userViewModel.deleteAllUsers()
-            Toast.makeText(requireContext().applicationContext, "All data successfully deleted", Toast.LENGTH_SHORT).show()
-        }
-        builder.setNegativeButton("No"){_, _ ->
+    private fun deleteAllUsers() {
+        userViewModel.readAllData.observe(this, Observer { user ->
 
-        }
-        builder.setTitle("Delete Everything?")
-        builder.setMessage("Do you want to delete all data?")
-        builder.create().show()
+            if (user.isNotEmpty()) {
+                //the database contains user(even a single one) delete all
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setPositiveButton("Yes") { _, _ ->
+                    userViewModel.deleteAllUsers()
+                    setupToast(requireContext(), "All data deleted successfully")
+                    //remove the observer after clicking yes to avoid the other condition being executed
+                    userViewModel.readAllData.removeObservers(this)
+                }
+                builder.setNegativeButton("No") { _, _ ->
+
+                }
+                builder.setTitle("Delete Everything?")
+                builder.setMessage("Do you want to delete all data?")
+                builder.create()
+                builder.show()
+
+
+            } else if (user.isEmpty()) {
+                //when there are no users just display this alert
+                val builder = AlertDialog.Builder(requireContext())
+                builder.setNegativeButton("Ok") { _, _ ->
+
+                }
+                builder.setTitle("Sorry!")
+                builder.setMessage("We couldn't find any data to delete")
+                builder.create()
+                builder.show()
+
+            }
+
+        })
+
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
     }
+
+
 }
